Remove the use of r7 for thumb2 assembly
The register r7 in thumb2 assembly instructions was used for numeric
operations. However, the compiler also uses r7 as a stack frame pointer.
Therefore, trying to compile a program without -fomit-frame-pointer
would fail and even passing this option there is no guarantee that r7
will not be used by the compiler. Therefore, the use of r7 was removed
altogether from the multiply assembly sequence in bn_mul.h